refactor: isolate FFI
