+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "crates/bls-crypto", 
     "crates/bls-gadgets",
     "crates/epoch-snark",
+    "crates/bls-snark-sys",
 ]
 
 [profile.release]
