+++ crates/bls-snark-sys/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "bls-snark-sys"
+version = "0.1.0"
+authors = ["Georgios Konstantopoulos <me@gakonst.com>"]
+edition = "2018"
+
+[dependencies]
+bls-crypto = { path = "../bls-crypto", features = ["compat"] }
+epoch-snark = { path = "../epoch-snark", features = ["compat"] }
+
+algebra = { git = "https://github.com/scipr-lab/zexe", default-features = false, features = ["bls12_377", "parallel"] }
+once_cell = "1.4.0"
+rand = "0.7.3"
+log = "0.4.8"
+
+[dev-dependencies]
+groth16 = { git = "https://github.com/scipr-lab/zexe", features = ["parallel"] }
+r1cs-core = { git = "https://github.com/scipr-lab/zexe" }
+hex = "0.4.2"
+
