+++ crates/bls-snark-sys/src/signatures.rs
@@ -1,10 +1,10 @@
-use super::{
+use crate::{
     convert_result_to_bool,
     utils::{Message, MessageFFI},
     PrivateKey, PublicKey, PublicKeyCache, Signature, COMPOSITE_HASH_TO_G1, DIRECT_HASH_TO_G1,
 };
-use crate::{BLSError, HashToCurve, POP_DOMAIN, SIG_DOMAIN};
 use algebra::{ProjectiveCurve, ToBytes};
+use bls_crypto::{BLSError, HashToCurve, POP_DOMAIN, SIG_DOMAIN};
 use std::{os::raw::c_int, slice};
 
 use once_cell::sync::Lazy;
