+++ crates/bls-snark-sys/src/utils.rs
