+++ crates/epoch-snark/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Kobi Gurkan <kobigurk@gmail.com>", "Michael Straka <mstraka@celo.org
 edition = "2018"
 
 [dependencies]
-bls-crypto = { path = "../bls-crypto", default-features = false, features = ["ffi"] }
+bls-crypto = { path = "../bls-crypto", default-features = false }
 bls-gadgets = { path = "../bls-gadgets", default-features = false }
 
 algebra = { git = "https://github.com/scipr-lab/zexe", features = ["bls12_377", "sw6", "edwards_sw6", "edwards_bls12", "parallel"] }
@@ -25,7 +25,6 @@ tracing = "0.1.13"
 [dev-dependencies]
 rand_xorshift = { version = "0.2" }
 bench-utils = { git = "https://github.com/scipr-lab/zexe" }
-hex = "0.3.2"
 bls-gadgets = { path = "../bls-gadgets", default-features = false, features = ["test-helpers"] }
 bls-crypto = { path = "../bls-crypto", default-features = false, features = ["test-helpers"] }
