+++ crates/epoch-snark/tests/e2e.rs
@@ -1,5 +1,4 @@
-use algebra::CanonicalSerialize;
-use epoch_snark::api::{ffi, prover, setup, verifier};
+use epoch_snark::api::{prover, setup, verifier};
 
 mod fixtures;
 use fixtures::generate_test_data;
@@ -35,47 +34,4 @@ fn prover_verifier_groth16() {
     // Verifier checks the proof
     let res = verifier::verify(&params.epochs.vk, &first_epoch, &last_epoch, &proof);
     assert!(res.is_ok());
-
-    // Serialize the proof / vk
-    let mut serialized_vk = vec![];
-    params.epochs.vk.serialize(&mut serialized_vk).unwrap();
-    let mut serialized_proof = vec![];
-    proof.serialize(&mut serialized_proof).unwrap();
-    dbg!(hex::encode(&serialized_vk));
-    dbg!(hex::encode(&serialized_proof));
-
-    // Get the corresponding pointers
-    let proof_ptr = &serialized_proof[0] as *const u8;
-    let vk_ptr = &serialized_vk[0] as *const u8;
-
-    let serialized_pubkeys = ffi::utils::serialize_pubkeys(&first_epoch.new_public_keys).unwrap();
-    dbg!(hex::encode(&serialized_pubkeys));
-    let first_epoch = ffi::utils::EpochBlockFFI {
-        index: first_epoch.index,
-        maximum_non_signers: first_epoch.maximum_non_signers,
-        pubkeys_num: first_epoch.new_public_keys.len(),
-        pubkeys: &serialized_pubkeys[0] as *const u8,
-    };
-
-    let serialized_pubkeys = ffi::utils::serialize_pubkeys(&last_epoch.new_public_keys).unwrap();
-    dbg!(hex::encode(&serialized_pubkeys));
-    let last_epoch = ffi::utils::EpochBlockFFI {
-        index: last_epoch.index,
-        maximum_non_signers: last_epoch.maximum_non_signers,
-        pubkeys_num: last_epoch.new_public_keys.len(),
-        pubkeys: &serialized_pubkeys[0] as *const u8,
-    };
-
-    // Make the verification
-    let res = unsafe {
-        ffi::verify(
-            vk_ptr,
-            serialized_vk.len() as u32,
-            proof_ptr,
-            serialized_proof.len() as u32,
-            first_epoch,
-            last_epoch,
-        )
-    };
-    assert!(res);
 }
